Internals

Everything documented in this page is internal and subject to breaking changes. These are provided here for those curious about how JuliaSyntaxHighlighting works, and as a reference for contributors.

JuliaSyntaxHighlighting.MAX_PAREN_HIGHLIGHT_DEPTH — Constant

MAX_PAREN_HIGHLIGHT_DEPTH

The number of julia_rainbow_{paren,bracket_curly}_{n} faces from HIGHLIGHT_FACES that can be cycled through.

JuliaSyntaxHighlighting.RAINBOW_DELIMITERS_ENABLED — Constant

RAINBOW_DELIMITERS_ENABLED

Whether to use julia_rainbow_{paren,bracket_curly}_{n} faces for delimitors/parentheses ((), [], {}) as opposed to just using julia_parentheses.

JuliaSyntaxHighlighting.UNMATCHED_DELIMITERS_ENABLED — Constant

UNMATCHED_DELIMITERS_ENABLED

Whether to apply the julia_unpaired_parentheses face to unpaired closing parenthesis (), ], '}').

JuliaSyntaxHighlighting.SINGLETON_IDENTIFIERS — Constant

SINGLETON_IDENTIFIERS

Symbols that represent identifiers known to be instances of a singleton type, currently just Nothing and Missing.

JuliaSyntaxHighlighting.BASE_TYPE_IDENTIFIERS — Constant

BASE_TYPE_IDENTIFIERS

A set of type identifiers defined in Base or Core.

JuliaSyntaxHighlighting.BUILTIN_FUNCTIONS — Constant

BUILTIN_FUNCTIONS

A set of identifiers that are defined in Core and a Core.Builtin.

JuliaSyntaxHighlighting.HIGHLIGHT_FACES — Constant

HIGHLIGHT_FACES

A list of name => Face(...) pairs that define the faces in JuliaSyntaxHighlighting. These are registered during module initialisation.

JuliaSyntaxHighlighting.paren_type — Function

paren_type(k::Kind) -> (Int, Symbol)

Return a pair of values giving the change in nesting depth caused by the paren k (+1 or -1), as well as a symbol indicating the kind of parenthesis:

( and ) are a paren

[ and ] are a bracket

{ and } are a curly

Anything else is of type none, and produced a depth change of 0.

JuliaSyntaxHighlighting._hl_annotations — Function

_hl_annotations(content::AbstractString, ast::GreenNode) -> Vector{Tuple{UnitRange{Int}, Pair{Symbol, Any}}}

Generate a list of (range, annot) pairs for the given content and ast.

The range is a UnitRange{Int} that indexes into ctx.content and annot is a Pair of the form :face => <face>.

This is a small wrapper around _hl_annotations! for convenience.

JuliaSyntaxHighlighting._hl_annotations! — Function

_hl_annotations!(highlights::Vector{Tuple{UnitRange{Int}, Pair{Symbol, Any}}},
                 lineage::GreenLineage, ctx::HighlightContext)

Populate highlights with (range, annot) pairs for the given lineage and ctx, where lineage is expected to be consistent with ctx.offset and ctx.lnode.

The range is a UnitRange{Int} that indexes into ctx.content and annot is a Pair of the form :face => <face>.